+++ pypicloud/access/__init__.py
@@ -23,7 +23,7 @@ def includeme(config):
     elif dotted_name == 'sql':
         dotted_name = SQLAccessBackend
     elif dotted_name == 'ldap':
-        dotted_name = "pypicloud.access._ldap.LDAPAccessBackend"
+        dotted_name = "pypicloud.access.ldap_.LDAPAccessBackend"
     access_backend = resolver.maybe_resolve(dotted_name)
     kwargs = access_backend.configure(settings)
     config.add_request_method(partial(access_backend, **kwargs), name='access',