+++ doc/changes.rst
@@ -5,6 +5,7 @@ If you are upgrading an existing installation, read :ref:`the instructions <upgr
 0.3.3
 -----
 * Feature: LDAP Support (:pr:`55`)
+* Bug fix: Incorrect package name/version when uploading from web (:issue:`56`)
 
 0.3.2
 -----